WSGILogger(..., ip_header='HTTP_X_FORWARDED_FOR', ...) kw arg
This new keyword argument allows to use headers such as X-Forwarded-For
for the 'host' field in your logging. This is useful if you run behind a
reverse proxy. You have to specify the header as it will be provided by
the WSGI environment, in this case it would be HTTP_X_FORWARDED_FOR.
All characters are in uppercase and dashes will be replaced by underscores.

If you don't specify the keyword argument, it defaults to None. Then, or
whenever the header is not found in the WSGI environment, the logged host
will default to the REMOTE_ADDR.